Conflicts:
	src/flowpath.cpp

+++ src/dtm.inl
@@ -24,11 +24,11 @@ template<Primitive prim> inline void DTM::build_begin() const
     }
 }
 
-template<Primitive prim> inline void DTM::build_line(unsigned long) {}
-template<Primitive prim> inline void DTM::build_line(unsigned long) const {}
+template<Primitive prim> inline void DTM::build_line(GLuint) {}
+template<Primitive prim> inline void DTM::build_line(GLuint) const {}
 
-template<Primitive prim> inline void DTM::build_back(unsigned long) {}
-template<Primitive prim> inline void DTM::build_back(unsigned long) const {}
+template<Primitive prim> inline void DTM::build_back(GLuint) {}
+template<Primitive prim> inline void DTM::build_back(GLuint) const {}
 
 template<Primitive prim> inline void DTM::build_end() const
 {
@@ -53,7 +53,7 @@ template<Primitive prim> inline void DTM::build_end() const
  * POINTS
  */
 
-TEMPLATE(build_line, POINT, unsigned long i, buildFunction(i);)
+TEMPLATE(build_line, POINT, GLuint i, buildFunction(i);)
 #define ARRAY_SIZE_POINT NB_CALL_LINE
 
 
@@ -62,20 +62,20 @@ TEMPLATE(build_line, POINT, unsigned long i, buildFunction(i);)
  * EDGES
  */
 
-TEMPLATE(build_line, TRILINE, unsigned long i,
+TEMPLATE(build_line, TRILINE, GLuint i,
          buildFunction(i);
          if(i < m_vertices.size() - m_lineLength) // not at last line
              buildFunction(i + m_lineLength);)
 
-TEMPLATE(build_back, TRILINE, unsigned long j, buildFunction(j);)
+TEMPLATE(build_back, TRILINE, GLuint j, buildFunction(j);)
 
 #define ARRAY_SIZE_TRILINE NB_CALL_LINE*2 - m_lineLength + NB_CALL_BACK
 
 
 
-TEMPLATE(build_line, QUADLINE, unsigned long i,  build_line<TRILINE>(i);)
+TEMPLATE(build_line, QUADLINE, GLuint i,  build_line<TRILINE>(i);)
 
-TEMPLATE(build_back, QUADLINE, unsigned long j,
+TEMPLATE(build_back, QUADLINE, GLuint j,
     if(j % m_lineLength) { // not at beginning of line
         buildFunction(j);
         buildFunction(j-1);
@@ -89,7 +89,7 @@ TEMPLATE(build_back, QUADLINE, unsigned long j,
  * SHAPES
  */
 
-TEMPLATE(build_line, TRIFILL, unsigned long i,
+TEMPLATE(build_line, TRIFILL, GLuint i,
     buildFunction(i);
     if(i < m_vertices.size() - m_lineLength) // not at last line
         buildFunction(i + m_lineLength);
@@ -99,7 +99,7 @@ TEMPLATE(build_line, TRIFILL, unsigned long i,
 
 #define ARRAY_SIZE_TRIFILL NB_CALL_LINE*2
 
-TEMPLATE(build_line, QUADFILL, unsigned long i, build_line<TRIFILL>(i);)
+TEMPLATE(build_line, QUADFILL, GLuint i, build_line<TRIFILL>(i);)
 
 #define ARRAY_SIZE_QUADFILL ARRAY_SIZE_TRIFILL
 
@@ -114,7 +114,7 @@ TEMPLATE(build_line, QUADFILL, unsigned long i, build_line<TRIFILL>(i);)
 template<> inline void DTM::build_begin<LINELOOP>() const {}
 template<> inline void DTM::build_end  <LINELOOP>() const {}
 
-TEMPLATE(build_line, LINELOOP, unsigned long i,
+TEMPLATE(build_line, LINELOOP, GLuint i,
     if(getMode() == MODE_LEGACY)
     {
         glBegin(GL_LINE_LOOP);
@@ -141,7 +141,7 @@ TEMPLATE(build_line, LINELOOP, unsigned long i,
  * TRIANGLES
  */
 
-TEMPLATE(build_line, TRIANGLES, unsigned long i,
+TEMPLATE(build_line, TRIANGLES, GLuint i,
     if(i % m_lineLength && i < m_vertices.size() - m_lineLength) { // not at beginning nor last line
         buildFunction(i-1);
         buildFunction(i-1 + m_lineLength);
